requirements.txt
pyrogram
tgcrypto
apscheduler
python-dotenv
motor
